解决前端路由404问题，打通趣事录与旅游志的图片，优化展示页面
