+++ frontend/src/components/CreateEvent.vue
@@ -1,52 +1,3 @@
-<!-- <template>
-    <div>
-        <h2>创建趣事录</h2>
-        <form @submit.prevent="submitForm">
-            <label for="name">趣事录名称:</label>
-            <input type="text" id="name" v-model="eventName" required>
-            <br>
-            <label for="photo">趣事录封面图:</label>
-            <input type="file" id="photo" @change="handlePhotoUpload" required>
-            <br>
-            <button type="submit">提交</button>
-        </form>
-    </div>
-</template>
-
-<script>
-export default {
-    data() {
-        return {
-            eventName: '',
-            eventphoto: null,
-        };
-    },
-    methods: {
-        handlePhotoUpload(event) {
-            this.photo = event.target.files[0];
-        },
-        submitForm() {
-            // Perform API call to send eventName and photo to the backend
-            // Example using axios:
-            const formData = new FormData();
-            formData.append('eventName', this.eventName);
-            formData.append('photo', this.eventphoto);
-            console.log(formData.value);
-
-            axios.post('/api/create-interesting-event', formData)
-                .then(response => {
-                    // Handle success response
-                    console.log(response.data);
-                })
-                .catch(error => {
-                    // Handle error response
-                    console.error(error);
-                });
-        }
-    }
-};
-</script> -->
-
 <template>
   <div class="title">创建{{ title }}</div>
   <el-form ref="ruleFormRef" style="max-width: 600px" :model="ruleForm" status-icon :rules="rules" label-width="auto"
@@ -57,15 +8,14 @@ export default {
     <el-form-item :label="title+'描述'" prop="album_description">
       <el-input type="textarea" v-model="ruleForm.album_description" autocomplete="off" />
     </el-form-item>
-    <el-form-item :label="title+'封面图'" prop="album_cover">
-       <!-- <el-input v-model="ruleForm.album_cover" autocomplete="off" /> -->
+    <!-- <el-form-item :label="title+'封面图'" prop="album_cover">
       <el-upload class="upload-demo" :show-file-list="false" :before-upload="beforeUpload" :on-success="handleUploadSuccess">
         <img v-if="ruleForm.album_cover" :src="ruleForm.album_cover" style="max-width: 200px;max-height: 200px;" alt="趣事录封面图" />
         <el-icon v-else class="avatar-uploader-icon">
           <Plus />
         </el-icon>
       </el-upload>
-    </el-form-item>
+    </el-form-item> -->
     <el-form-item>
       <el-button type="primary" @click="submitForm()">提交</el-button>
       <el-button @click="resetForm()">取消</el-button>
@@ -100,32 +50,30 @@ export default{
     }
   },
   methods:{
-    beforeUpload(file){
-      console.log('file',file);
-      console.log('开始');
-      const url=URL.createObjectURL(file);//将文件转为url
-      this.ruleForm.album_cover=url;
-      console.log('url',this.ruleForm.album_cover);
-      return false;//阻止默认上传
-    },
-    handleUploadSuccess(response,file){
-      if(response.code === 200){
-        this.ruleForm.album_cover = response.data.url;
-        console.log('上传成功',response.data.url);
-      }else{
-        this.$message.error('上传失败');
-      }  
-    },
+    // beforeUpload(file){
+    //   console.log('file',file);
+    //   console.log('开始');
+    //   const url=URL.createObjectURL(file);//将文件转为url
+    //   this.ruleForm.album_cover=url;
+    //   console.log('url',this.ruleForm.album_cover);
+    //   return false;//阻止默认上传
+    // },
+    // handleUploadSuccess(response,file){
+    //   if(response.code === 200){
+    //     this.ruleForm.album_cover = response.data.url;
+    //     console.log('上传成功',response.data.url);
+    //   }else{
+    //     this.$message.error('上传失败');
+    //   }  
+    // },
     submitForm(){
       this.$refs.ruleFormRef.validate((valid)=>{
         if(valid){
-          console.log('提交成功!');
           if(this.title === '趣事录'){
             this.ruleForm.album_type = 2;
           }else if(this.title === '旅游志'){
             this.ruleForm.album_type = 3;
           }
-          console.log(this.ruleForm);
           //发送请求
           axios.post('/album',this.ruleForm).then(response=>{
             console.log(response.data);
