+++ frontend/src/views/CreateClassmate.vue
@@ -31,10 +31,9 @@ export default{
   },
   methods:{
     handleFormSubmit(newData){
-      console.log('add里');
-      console.log(this.formData);
       this.formData=newData;
-      console.log('formdata:',this.formData);
+      const type = this.$router.currentRoute.value.query.type;
+      this.formData.classmates_avatar_name = type;
       /* 如果表单不完整，则返回；否则，则将数据发送给数据库 */
       if (Object.values(this.formData).some(value => !value))
       ElMessageBox.alert('请填写完整信息', '提示', {
@@ -45,9 +44,7 @@ export default{
       axios.post('/classmate/', this.formData)
         .then(response => {
           // 处理成功响应
-          console.log(response.data);
           this.formData={};//清空表单,但失败???
-          console.log('formdata:',this.formData);
           ElMessageBox.alert('添加成功', '提示', {
             confirmButtonText: '确定',
             type: 'success'
@@ -56,8 +53,8 @@ export default{
               confirmButtonText: '确定',
               type: 'info'
             });
-            // const type = this.$router.currentRoute.value.query.type;
-            // this.$router.push({ path: '/classmates/informshow',query: { stage: type } });
+            const type = this.$router.currentRoute.value.query.type;
+            this.$router.push({ path: '/classmates/informshow',query: { stage: type } });
           });
         })
         .catch(error => {
